Find root scene from viewport configuration
Find the root scene from viewport configuration so that we are killing the
minimum number of scenes required. Otherwise if the tree of scenes was deep we
might kill the parent scene and then try to kill the child scene which would
result in an error. With this change we only kill the root scene which will
correctly restart all of the children.